flask_chat_app/requirements.txt
gevent==1.4.0
flask==1.1.2
gevent-websocket
flask_cors==3.0.10
gunicorn==20.1.0
flask-socketio
eventlet==0.30.2
redis==3.5.3
psycopg2-binary==2.8.6
netifaces==0.10.4
requests==2.18.4
pyyaml==3.12
